+++ src/naming-strategies.ts
